Changed the Readme to include a small example of how to get notifio working inside a python script.
I searched the internet for hours trying to figure out
how to get this module to work after installing it.

I ended up reading to source of another program that had used
notifo and figured it out.

Having to "from notifo import Notifo" to get things working
maked it hard to hit the ground running.

The examples that were originally included all threw errors for me.